2022 day 22 p2 finished
Refactoring and more testing
Found that Wall detection was not working properly. This was because we didn't update the Map back to old locatin, when hitting a wall.

I do like these kinds of tasks, but this was a pain. 2d and 3d pixel movement stuff is cool and fun. Had to make a paper cube to see the faces and rotations and figure out how to map the given input to a cube.